Release Wasmtime 17.0.0 (#7800)
* Release Wasmtime 17.0.0

[automatically-tag-and-release-this-commit]

* Change update of gcc on MinGW (#7760)

Try not passing `-y -u` to `pacman` to avoid full system updates.
Currently full system updates might update the `msys2-runtime` package
before actually updating the package we requested, meaning that this
might not actually update anything given an update. This is what's
currently happening on CI which is breaking due to an update of gcc not
actually updating gcc. I'm mostly reading the invocation in
rust-lang/rust CI and seeing that it doesn't pass `-y -u` and hopeing
that by copying that here things might work.

prtest:full

* use released WASI Preview 2 (version 0.2.0) wits (#7817)

* WASI: copy in the version 0.2.0 wits

* wasmtime's wits: use versions 0.2.0 of wasi packages

* bindgens and other fixed version strings: change 0.2.0-rc-etc to 0.2.0

* Enable the component model by default

This commit enables the component model by default in the embedding API
and the CLI. This means that an opt-in of `-W component-model` is no
longer required and additionally `.wasm_component_model(true)` is no
longer required. Note that this won't impact existing embeddings since
the component model feature doesn't do much less `wasmtime::component`
is used, and if that's being used this is probably good news for you.

* Add release notes for 17.0.0

* Fix non-component-model build

---------

Co-authored-by: Wasmtime Publish <wasmtime-publish@users.noreply.github.com>
Co-authored-by: Alex Crichton <alex@alexcrichton.com>
Co-authored-by: Pat Hickey <phickey@fastly.com>