Disallow values from jumping between blocks inappropriately
Undo changes to the union find structure and gvn_map when popping
the node stack in `remove_pure_and_optimize`, ensuring that we don't
propagate values in a way that would invalidate the invariant that the
`ValueData::Union` structure provides.

Additionally, remove all uses of the union find structure from
elaboration, as we have already computed the best values at that point.

Co-authored-by: Jamey Sharp <jsharp@fastly.com>
Co-authored-by: L Pereira <l.pereira@fastly.com>
Co-authored-by: Nick Fitzgerald <fitzgen@gmail.com>