Replace WASI WITs with binary form
This commit replaces the textual `*.wit` files in the Wasmtime tree with
the binary form of a WIT package. The new WIT structure now looks like:

* `crates/wasi/wasi-cli@0.2.0.wasm` - serialized version of the `wasi-cli`
  proposal at the v0.2.0 tag. Created with `wasm-tools component wit
  --wasm`

* `crates/wasi-http/wasi-http@0.2.0.wasm` - same as the above, but for
  `wasi-http`.

* `crates/test-programs/wit` - WIT files used for tests such as a custom
  reactor.

* `crates/test-programs/wit/deps/*.wasm` - WIT bindings to generate
  tests with. This is a duplicate of the `wasi-{cli,http}@0.2.0.wasm`
  files from above but additionally enables testing multiple versions in
  the future.

The end state of this PR is that whenever WASI APIs are updated we'll
need a tagged release to pull into Wasmtime to officially support.