cranelift: Remove redundant dominator-tree structure (#7948)
* Don't keep extra copies of root block ID

We already know the entry block everywhere we use this.

* Replace DomTreeWithChildren with DominatorTreePreorder

These two types had nearly identical interfaces except that the latter
also supports constant-time dominance checks.

* Use dom-tree preorder for O(1) dominance checks

It turns out this was the only thing the egraph pass used the original
dominator tree for, so we can stop passing that around.

We could also require that the caller of EgraphPass::new construct the
DominatorTreePreorder and avoid passing the original tree into the
egraph pass at all, but I've chosen to stop refactoring here.

We should review all uses of DominatorTree::dominates to see if it makes
sense to pass a DominatorTreePreorder around more places.

* review comments: Rename/document away from "domtree with children"