Enable compiling the Wasmtime CLI to Wasm (#7980)
* Enable compiling the Wasmtime CLI to Wasm

While not the most useful thing to do in the world it's kind of neat to
play around with. This builds on the previous work to exclude the
runtime from the `wasmtime` crate so it's now possible to compile the
Wasmtime CLI's `compile` command, and only the `compile` command, to
wasm itself. This means you can run Wasmtime in Wasmtime!

* Fix warning on wasm

* Fix some feature combos